+++ package.json
@@ -25,8 +25,8 @@
 	},
 	"devDependencies": {
 		"@playwright/test": "^1.17.1",
-		"@sveltejs/kit": "^1.0.0-next.437",
-		"@sveltejs/package": "^1.0.0-next.1",
+		"@sveltejs/kit": "^1.0.0-next.480",
+		"@sveltejs/package": "^1.0.0-next.3",
 		"@types/node": "^17.0.0",
 		"@types/prettier": "^2.4.3",
 		"cross-env": "^7.0.3",
