+++ src/hooks.server.ts
@@ -1,0 +1,6 @@
+import { handleSession } from '$lib/session';
+
+export const handle = handleSession({
+	key: 'svemix.testing',
+	secret: '123456789101112'
+});
