+++ src/routes/metadata/index.svelte
@@ -1,4 +1,4 @@
 <script context="module" ssr lang="ts">
-	import type { MetaFn } from '$lib/server';
+	import type { MetaFn } from '$lib/core';
 	export const metadata: MetaFn = () => ({ title: 'Custom Title' });
 </script>
